+++ spring-data-jdbc/src/main/java/org/springframework/data/jdbc/repository/config/AbstractJdbcConfiguration.java
@@ -135,8 +135,7 @@ public class AbstractJdbcConfiguration implements ApplicationContextAware {
 				: JdbcArrayColumns.DefaultSupport.INSTANCE;
 		DefaultJdbcTypeFactory jdbcTypeFactory = new DefaultJdbcTypeFactory(operations.getJdbcOperations(), arrayColumns);
 
-		return new BasicJdbcConverter(mappingContext, relationResolver, conversions, jdbcTypeFactory,
-				dialect.getIdentifierProcessing());
+		return new MappingJdbcConverter(mappingContext, relationResolver, conversions, jdbcTypeFactory);
 	}
 
 	/**
