+++ spring-data-jdbc/src/test/java/org/springframework/data/jdbc/core/convert/DefaultDataAccessStrategyUnitTests.java
@@ -45,11 +45,11 @@ class DefaultDataAccessStrategyUnitTests {
 
 	static final long ORIGINAL_ID = 4711L;
 
-	private NamedParameterJdbcOperations namedJdbcOperations = mock(NamedParameterJdbcOperations.class);
-	private JdbcOperations jdbcOperations = mock(JdbcOperations.class);
-	private RelationalMappingContext context = new JdbcMappingContext();
-	private SqlParametersFactory sqlParametersFactory = mock(SqlParametersFactory.class);
-	private InsertStrategyFactory insertStrategyFactory = mock(InsertStrategyFactory.class);
+	private final NamedParameterJdbcOperations namedJdbcOperations = mock(NamedParameterJdbcOperations.class);
+	private final JdbcOperations jdbcOperations = mock(JdbcOperations.class);
+	private final RelationalMappingContext context = new JdbcMappingContext();
+	private final SqlParametersFactory sqlParametersFactory = mock(SqlParametersFactory.class);
+	private final InsertStrategyFactory insertStrategyFactory = mock(InsertStrategyFactory.class);
 
 	private JdbcConverter converter;
 	private DataAccessStrategy accessStrategy;
@@ -59,7 +59,7 @@ class DefaultDataAccessStrategyUnitTests {
 
 		DelegatingDataAccessStrategy relationResolver = new DelegatingDataAccessStrategy();
 		Dialect dialect = HsqlDbDialect.INSTANCE;
-		converter = new BasicJdbcConverter(context, relationResolver, new JdbcCustomConversions(),
+		converter = new MappingJdbcConverter(context, relationResolver, new JdbcCustomConversions(),
 				new DefaultJdbcTypeFactory(jdbcOperations), dialect.getIdentifierProcessing());
 		accessStrategy = new DataAccessStrategyFactory( //
 				new SqlGeneratorSource(context, converter, dialect), //
