+++ spring-data-jdbc/src/test/java/org/springframework/data/jdbc/core/convert/SqlGeneratorFixedNamingStrategyUnitTests.java
@@ -19,7 +19,6 @@ import static org.assertj.core.api.Assertions.*;
 
 import org.assertj.core.api.SoftAssertions;
 import org.junit.jupiter.api.Test;
-
 import org.springframework.data.annotation.Id;
 import org.springframework.data.jdbc.core.PersistentPropertyPathTestUtils;
 import org.springframework.data.jdbc.core.mapping.JdbcMappingContext;
@@ -197,7 +196,7 @@ class SqlGeneratorFixedNamingStrategyUnitTests {
 	private SqlGenerator configureSqlGenerator(NamingStrategy namingStrategy) {
 
 		context = new JdbcMappingContext(namingStrategy);
-		JdbcConverter converter = new BasicJdbcConverter(context, (identifier, path) -> {
+		JdbcConverter converter = new MappingJdbcConverter(context, (identifier, path) -> {
 			throw new UnsupportedOperationException();
 		});
 		RelationalPersistentEntity<?> persistentEntity = context.getRequiredPersistentEntity(DummyEntity.class);
