+++ spring-data-relational/src/main/java/org/springframework/data/relational/core/conversion/MappingRelationalConverter.java
@@ -15,18 +15,23 @@
  */
 package org.springframework.data.relational.core.conversion;
 
+import java.lang.reflect.Array;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.Optional;
+import java.util.function.Function;
 import java.util.function.Predicate;
 
 import org.springframework.beans.BeansException;
 import org.springframework.context.ApplicationContext;
 import org.springframework.context.ApplicationContextAware;
 import org.springframework.core.CollectionFactory;
+import org.springframework.core.ResolvableType;
 import org.springframework.core.convert.ConversionService;
+import org.springframework.core.convert.TypeDescriptor;
 import org.springframework.data.convert.CustomConversions;
 import org.springframework.data.mapping.InstanceCreatorMetadata;
 import org.springframework.data.mapping.MappingException;
@@ -34,6 +39,7 @@ import org.springframework.data.mapping.Parameter;
 import org.springframework.data.mapping.PersistentEntity;
 import org.springframework.data.mapping.PersistentProperty;
 import org.springframework.data.mapping.PersistentPropertyAccessor;
+import org.springframework.data.mapping.PersistentPropertyPathAccessor;
 import org.springframework.data.mapping.context.MappingContext;
 import org.springframework.data.mapping.model.ConvertingPropertyAccessor;
 import org.springframework.data.mapping.model.DefaultSpELExpressionEvaluator;
@@ -41,6 +47,7 @@ import org.springframework.data.mapping.model.EntityInstantiator;
 import org.springframework.data.mapping.model.ParameterValueProvider;
 import org.springframework.data.mapping.model.PersistentEntityParameterValueProvider;
 import org.springframework.data.mapping.model.PropertyValueProvider;
+import org.springframework.data.mapping.model.SimpleTypeHolder;
 import org.springframework.data.mapping.model.SpELContext;
 import org.springframework.data.mapping.model.SpELExpressionEvaluator;
 import org.springframework.data.mapping.model.SpELExpressionParameterValueProvider;
@@ -69,9 +76,15 @@ import org.springframework.util.ClassUtils;
  * {@link RowDocument}.
  *
  * @author Mark Paluch
+ * @author Jens Schauder
+ * @author Chirag Tailor
+ * @author Vincent Galloy
+ * @see MappingContext
+ * @see SimpleTypeHolder
+ * @see CustomConversions
  * @since 3.2
  */
-public class MappingRelationalConverter extends BasicRelationalConverter implements ApplicationContextAware {
+public class MappingRelationalConverter extends AbstractRelationalConverter implements ApplicationContextAware {
 
 	private SpELContext spELContext;
 
@@ -124,11 +137,6 @@ public class MappingRelationalConverter extends BasicRelationalConverter impleme
 		this.projectionFactory.setBeanClassLoader(applicationContext.getClassLoader());
 	}
 
-	@Override
-	public ProjectionFactory getProjectionFactory() {
-		return this.projectionFactory;
-	}
-
 	/**
 	 * Creates a new {@link ConversionContext}.
 	 *
@@ -140,6 +148,14 @@ public class MappingRelationalConverter extends BasicRelationalConverter impleme
 
 		return new DefaultConversionContext(this, getConversions(), path, this::readAggregate, this::readCollectionOrArray,
 				this::readMap, this::getPotentiallyConvertedSimpleRead);
+	}
+
+	@Override
+	public <T> PersistentPropertyPathAccessor<T> getPropertyAccessor(PersistentEntity<T, ?> persistentEntity,
+			T instance) {
+
+		PersistentPropertyPathAccessor<T> accessor = persistentEntity.getPropertyPathAccessor(instance);
+		return new ConvertingPropertyAccessor<>(accessor, getConversionService());
 	}
 
 	@Override
@@ -315,7 +331,7 @@ public class MappingRelationalConverter extends BasicRelationalConverter impleme
 		}
 
 		if (RowDocument.class.isAssignableFrom(rawType)) {
-			return (S) documentAccessor;
+			return (S) documentAccessor.document();
 		}
 
 		if (typeHint.isMap()) {
@@ -445,6 +461,14 @@ public class MappingRelationalConverter extends BasicRelationalConverter impleme
 		return instance;
 	}
 
+	@Override
+	public <T> T createInstance(PersistentEntity<T, RelationalPersistentProperty> entity,
+			Function<Parameter<?, RelationalPersistentProperty>, Object> parameterValueProvider) {
+
+		return getEntityInstantiators().getInstantiatorFor(entity) //
+				.createInstance(entity, new ConvertingParameterValueProvider<>(parameterValueProvider));
+	}
+
 	private ParameterValueProvider<RelationalPersistentProperty> getParameterProvider(ConversionContext context,
 			RelationalPersistentEntity<?> entity, RowDocumentAccessor source, SpELExpressionEvaluator evaluator) {
 
@@ -566,6 +590,165 @@ public class MappingRelationalConverter extends BasicRelationalConverter impleme
 		}
 
 		return false;
+	}
+
+	@Override
+	@Nullable
+	public Object readValue(@Nullable Object value, TypeInformation<?> type) {
+
+		if (null == value) {
+			return null;
+		}
+
+		if (getConversions().hasCustomReadTarget(value.getClass(), type.getType())) {
+
+			TypeDescriptor sourceDescriptor = TypeDescriptor.valueOf(value.getClass());
+			TypeDescriptor targetDescriptor = createTypeDescriptor(type);
+
+			return getConversionService().convert(value, sourceDescriptor, targetDescriptor);
+		}
+
+		return getPotentiallyConvertedSimpleRead(value, type);
+	}
+
+	/**
+	 * Checks whether we have a custom conversion registered for the given value into an arbitrary simple JDBC type.
+	 * Returns the converted value if so. If not, we perform special enum handling or simply return the value as is.
+	 *
+	 * @param value to be converted. Must not be {@code null}.
+	 * @return the converted value if a conversion applies or the original value. Might return {@code null}.
+	 */
+	@Nullable
+	private Object getPotentiallyConvertedSimpleWrite(Object value) {
+
+		Optional<Class<?>> customTarget = getConversions().getCustomWriteTarget(value.getClass());
+
+		if (customTarget.isPresent()) {
+			return getConversionService().convert(value, customTarget.get());
+		}
+
+		return Enum.class.isAssignableFrom(value.getClass()) ? ((Enum<?>) value).name() : value;
+	}
+
+	/**
+	 * Checks whether we have a custom conversion for the given simple object. Converts the given value if so, applies
+	 * {@link Enum} handling or returns the value as is.
+	 *
+	 * @param value to be converted. May be {@code null}..
+	 * @param type {@link TypeInformation} into which the value is to be converted. Must not be {@code null}.
+	 * @return the converted value if a conversion applies or the original value. Might return {@code null}.
+	 */
+	@SuppressWarnings({ "rawtypes", "unchecked" })
+	protected Object getPotentiallyConvertedSimpleRead(Object value, TypeInformation<?> type) {
+
+		Class<?> target = type.getType();
+		if (ClassUtils.isAssignableValue(target, value)) {
+			return value;
+		}
+
+		if (Enum.class.isAssignableFrom(target) && value instanceof CharSequence) {
+			return Enum.valueOf((Class<Enum>) target, value.toString());
+		}
+
+		return getConversionService().convert(value, TypeDescriptor.forObject(value), createTypeDescriptor(type));
+	}
+
+	private static TypeDescriptor createTypeDescriptor(TypeInformation<?> type) {
+
+		List<TypeInformation<?>> typeArguments = type.getTypeArguments();
+		Class<?>[] generics = new Class[typeArguments.size()];
+		for (int i = 0; i < typeArguments.size(); i++) {
+			generics[i] = typeArguments.get(i).getType();
+		}
+
+		return new TypeDescriptor(ResolvableType.forClassWithGenerics(type.getType(), generics), type.getType(), null);
+	}
+
+	@Override
+	@Nullable
+	public Object writeValue(@Nullable Object value, TypeInformation<?> type) {
+
+		if (value == null) {
+			return null;
+		}
+
+		if (getConversions().isSimpleType(value.getClass())) {
+
+			if (TypeInformation.OBJECT != type && getConversionService().canConvert(value.getClass(), type.getType())) {
+				value = getConversionService().convert(value, type.getType());
+			}
+
+			return getPotentiallyConvertedSimpleWrite(value);
+		}
+
+		if (value.getClass().isArray()) {
+			return writeArray(value, type);
+		}
+
+		if (value instanceof Collection<?>) {
+			return writeCollection((Iterable<?>) value, type);
+		}
+
+		RelationalPersistentEntity<?> persistentEntity = getMappingContext().getPersistentEntity(value.getClass());
+
+		if (persistentEntity != null) {
+
+			Object id = persistentEntity.getIdentifierAccessor(value).getIdentifier();
+			return writeValue(id, type);
+		}
+
+		return getConversionService().convert(value, type.getType());
+	}
+
+	private Object writeArray(Object value, TypeInformation<?> type) {
+
+		Class<?> componentType = value.getClass().getComponentType();
+		Optional<Class<?>> optionalWriteTarget = getConversions().getCustomWriteTarget(componentType);
+
+		if (optionalWriteTarget.isEmpty() && !componentType.isEnum()) {
+			return value;
+		}
+
+		Class<?> customWriteTarget = optionalWriteTarget
+				.orElseGet(() -> componentType.isEnum() ? String.class : componentType);
+
+		// optimization: bypass identity conversion
+		if (customWriteTarget.equals(componentType)) {
+			return value;
+		}
+
+		TypeInformation<?> component = TypeInformation.OBJECT;
+		if (type.isCollectionLike() && type.getActualType() != null) {
+			component = type.getRequiredComponentType();
+		}
+
+		int length = Array.getLength(value);
+		Object target = Array.newInstance(customWriteTarget, length);
+		for (int i = 0; i < length; i++) {
+			Array.set(target, i, writeValue(Array.get(value, i), component));
+		}
+
+		return target;
+	}
+
+	private Object writeCollection(Iterable<?> value, TypeInformation<?> type) {
+
+		List<Object> mapped = new ArrayList<>();
+
+		TypeInformation<?> component = TypeInformation.OBJECT;
+		if (type.isCollectionLike() && type.getActualType() != null) {
+			component = type.getRequiredComponentType();
+		}
+
+		for (Object o : value) {
+			mapped.add(writeValue(o, component));
+		}
+
+		if (type.getType().isInstance(mapped) || !type.isCollectionLike()) {
+			return mapped;
+		}
+
+		return getConversionService().convert(mapped, type.getType());
 	}
 
 	static Predicate<RelationalPersistentProperty> isConstructorArgument(PersistentEntity<?, ?> entity) {
@@ -805,13 +988,16 @@ public class MappingRelationalConverter extends BasicRelationalConverter impleme
 		}
 	}
 
-	// TODO: Docs
+	/**
+	 * Extended {@link ParameterValueProvider} that can report whether a property value is present and contextualize the
+	 * instance for specific behavior like projection mapping in the context of a property.
+	 */
 	protected interface RelationalPropertyValueProvider extends PropertyValueProvider<RelationalPersistentProperty> {
 
 		/**
 		 * Determine whether there is a value for the given {@link RelationalPersistentProperty}.
 		 *
-		 * @param property
+		 * @param property the property to check for whether a value is present.
 		 * @return
 		 */
 		boolean hasValue(RelationalPersistentProperty property);
@@ -819,7 +1005,7 @@ public class MappingRelationalConverter extends BasicRelationalConverter impleme
 		/**
 		 * Contextualize this property value provider.
 		 *
-		 * @param context
+		 * @param context the context to use.
 		 * @return
 		 */
 		RelationalPropertyValueProvider withContext(ConversionContext context);
@@ -834,15 +1020,21 @@ public class MappingRelationalConverter extends BasicRelationalConverter impleme
 		/**
 		 * Determine whether there is a value for the given {@link AggregatePath}.
 		 *
-		 * @param path
+		 * @param path the path to check for whether a value is present.
 		 * @return
 		 */
 		boolean hasValue(AggregatePath path);
 
+		/**
+		 * Determine whether there is a value for the given {@link SqlIdentifier}.
+		 *
+		 * @param identifier the path to check for whether a value is present.
+		 * @return
+		 */
 		boolean hasValue(SqlIdentifier identifier);
 
 		/**
-		 * Returns a value for the given {@link AggregatePath}.
+		 * Return a value for the given {@link AggregatePath}.
 		 *
 		 * @param path will never be {@literal null}.
 		 * @return
@@ -868,9 +1060,13 @@ public class MappingRelationalConverter extends BasicRelationalConverter impleme
 	 * @author Mark Paluch
 	 * @author Christoph Strobl
 	 */
-	protected record DocumentValueProvider(ConversionContext context, RowDocumentAccessor accessor,
-			SpELExpressionEvaluator evaluator,
-			SpELContext spELContext) implements RelationalPropertyValueProvider, AggregatePathValueProvider {
+	protected static final class DocumentValueProvider
+			implements RelationalPropertyValueProvider, AggregatePathValueProvider {
+
+		private final ConversionContext context;
+		private final RowDocumentAccessor accessor;
+		private final SpELExpressionEvaluator evaluator;
+		private final SpELContext spELContext;
 
 		/**
 		 * Creates a new {@link RelationalPropertyValueProvider} for the given source and {@link SpELExpressionEvaluator}.
@@ -879,11 +1075,16 @@ public class MappingRelationalConverter extends BasicRelationalConverter impleme
 		 * @param accessor must not be {@literal null}.
 		 * @param evaluator must not be {@literal null}.
 		 */
-		protected DocumentValueProvider {
+		private DocumentValueProvider(ConversionContext context, RowDocumentAccessor accessor,
+				SpELExpressionEvaluator evaluator, SpELContext spELContext) {
 
 			Assert.notNull(context, "ConversionContext must no be null");
 			Assert.notNull(accessor, "DocumentAccessor must no be null");
 			Assert.notNull(evaluator, "SpELExpressionEvaluator must not be null");
+			this.context = context;
+			this.accessor = accessor;
+			this.evaluator = evaluator;
+			this.spELContext = spELContext;
 		}
 
 		@Override
@@ -936,6 +1137,38 @@ public class MappingRelationalConverter extends BasicRelationalConverter impleme
 			return context == this.context ? this : new DocumentValueProvider(context, accessor, evaluator, spELContext);
 		}
 
+		public ConversionContext context() {
+			return context;
+		}
+
+		public RowDocumentAccessor accessor() {
+			return accessor;
+		}
+
+	}
+
+	/**
+	 * Converter-aware {@link ParameterValueProvider}.
+	 *
+	 * @param <P>
+	 * @author Mark Paluch
+	 */
+	class ConvertingParameterValueProvider<P extends PersistentProperty<P>> implements ParameterValueProvider<P> {
+
+		private final Function<Parameter<?, P>, Object> delegate;
+
+		ConvertingParameterValueProvider(Function<Parameter<?, P>, Object> delegate) {
+
+			Assert.notNull(delegate, "Delegate must not be null");
+
+			this.delegate = delegate;
+		}
+
+		@Override
+		@SuppressWarnings("unchecked")
+		public <T> T getParameterValue(Parameter<T, P> parameter) {
+			return (T) readValue(delegate.apply(parameter), parameter.getType());
+		}
 	}
 
 	/**
